Fix names of some reference functions
The `__name__` field of some binary reference functions was wrong. We
fix this to be consistent with unary reference functions. In the future,
we should probably make the binary reference wrapper return a wrapper
itself to avoid all those calls to `partial`.

This change helps performing some homogeneous treatment of functions by
their name.

ghstack-source-id: cfb874fde782654cf05cc6d44c4a7607f3d8cb15
Pull Request resolved: https://github.com/pytorch/pytorch/pull/88115